chore(su) simplify the su code
