+++ servers/su/src/domain/mod.rs
@@ -1,15 +1,6 @@
 
-mod pipeline;
+
 mod clients;
 mod core;
 
-use clients::uploader::{UploaderClient};
-use clients::store::{StoreClient};
-
-// TODO: name these to be less confusing
-pub use pipeline::message::{MessagePipeline};
-pub use pipeline::messages::{MessagesPipeline};
-pub use pipeline::read_message::{ReadMessagePipeline};
-pub use pipeline::process::{ProcessPipeline};
-pub use pipeline::read_process::{ReadProcessPipeline};
-pub use pipeline::timestamp::{TimestampPipeline};
+pub mod flows;
