--- servers/su/src/domain/pipeline/mod.rs
@@ -1,7 +1,0 @@
-
-pub mod message;
-pub mod messages;
-pub mod read_message;
-pub mod process;
-pub mod read_process;
-pub mod timestamp;
